fixing inline server issue
